+++ tests/wasm/main.rs
@@ -56,6 +56,7 @@ pub mod result_jserror;
 pub mod rethrow;
 pub mod simple;
 pub mod slice;
+pub mod slice_jsvalue;
 pub mod string_vecs;
 pub mod struct_vecs;
 pub mod structural;